Creating Symlink Only If Its Raspberry Pi OS
According to this issue https://github.com/Botspot/pi-apps/issues/272 There is no need to create a symlink if its twister os. So I modified to code to only create a symlink if its raspberry pi os. I am checking if the os is twister os by checking if the binary twistver is there as its builtin in twiseros