Improve xlunch themes in light of recent fixes
Until recently, Xlunch would display png images without accounting for
any window transparency. This would "punch a hole" in the xlunch-dark
window, so to fix it I added my own semi-transparent background to the
images.
Now Xlunch has fixed this, so the transparent black background needs to
be removed from the images.
I've added a runonce entry to recompile xlunch, and removed pkexec
because it's annoying for passwordless-sudo people.