move python script used to set text color
botspots previous location caused the python script to run every time the preload script was called (aka, any time the page was changed the gui). this significantly slowed down the gui.

place the script at the top of the GUI and do not run it unless yad is older that 0.40 and text_color is not already available. this guarantees that the script will only run once during the entire use of pi-apps