Botspot's GUI update [DRAFT] (#1580)
* initial upload of new-gui changes

* Update gui

Co-authored-by: theofficialgman <28281419+theofficialgman@users.noreply.github.com>

* Update gui

Co-authored-by: theofficialgman <28281419+theofficialgman@users.noreply.github.com>

* Update gui

Co-authored-by: theofficialgman <28281419+theofficialgman@users.noreply.github.com>

* remove runonce for creating data/categories

* fix package-app install/uninstall

* preload: use colored app-name to indicate status

* gui: use colored app-name to indicate status

* undo argument input change

* revert

manually specify color and column to print

(for testing, will be reverted)

remove white hack and use printf

Update preload

* restructure yad so color is given at the end

workaround for bug seen: https://github.com/Botspot/pi-apps/issues/1578#issuecomment-1063647137

* correct missing color data

* disable logging

* add announcements back to GUI

* add mini python script to get theme text color

fixes issue: https://github.com/Botspot/pi-apps/issues/1578#issuecomment-1063657679

* workaround for yad < 0.40.0

only set default text_color for yad older than 0.40.0

* fix bug so ondemand preload generates prefix correctly

previously, only the preload-dameon correctly generated the InternetBrowsers list. this is because when the ondemand preloader went to generate the list, it did not have the required / inbetween prefix and dir to generate

* bugfix: add folders required for updater

if user has never run the manage script or did not do the install (like in testing) these required folders never get made and the updater script fails

* add runonce to clear data/preload after PR merge

* add error function

* Add wmctrl,xdotool,shellcheck,locales dependencies

* Focus app details window so it will close when unfocused

* workaround for xwayland display which does not contain "primary" in xrandr output

* force YAD and other windows to use xwayland

fixes bug with yad windows showing on random locations on screen when using a wayland based DE

* speed up 3-month-updater

* correct minimum yad version and use fast version method

* add initial manage script running check

Check if another instance of manage is running before starting. This prevents multiple install/uninstalls executed by the GUI from happening at once.

See commit contents for additional notes

* implement waitlist for apps

both the presence of a running manage script as well as any other app at the top of the waitlist will prevent an install from continuing

note: a check needs to be added to either continue or exit if the waitlist is somehow empty (like if an app was attempted to be installed twice)

* break out of waiting loop if waitlist is empty or non-existant

fallback incase the waitlist somehow is emptied or removed. Avoids an infinite loop.

* Update api

* preload-daemon: set text_color

* Revert "break out of waiting loop if waitlist is empty or non-existant"

This reverts commit dd3e1f69bb9cd25ad22c130dcfb7195d0bdab3fa.

* Revert "implement waitlist for apps"

This reverts commit 070b383c528732ffe904b8094cafd2f56003433c.

* Revert "add initial manage script running check"

This reverts commit 466107baf74970591a7ec2adea864afbed4128fc.

* gui: use manage daemon mode

* preload: set text_color

* manage: add new daemon mode

* icon changes

* manage: fix empty APT Packages files, fix multi-install

And prepare for some new icons

* Icon changes

* make api, gui, and install script executable again

* move python script used to set text color

botspots previous location caused the python script to run every time the preload script was called (aka, any time the page was changed the gui). this significantly slowed down the gui.

place the script at the top of the GUI and do not run it unless yad is older that 0.40 and text_color is not already available. this guarantees that the script will only run once during the entire use of pi-apps

* correctly interpret text_color blank set variable

* manage: only keep 6-day old logfiles to solve naming bug

* Rewrite logviewer for new icons & faster launch-time

* remove uninstalled.png from app list

* add settings tooltip to gui

* use GUI yad prompt to get user password if gui is available

if GUI is not available (aka, there is no display variable set), then fallback to getting the users password via the terminal (which should be available)

* edit dialog text

indicate why this popup is occuring

* (disabled) kill old GUI on update to the new GUI PR

comment out the lines made in this commit to enable this function

* Various improvements to main install script

* Fix spelling mistakes within the root of the directory (#1717)

* Fix spelling mistakes within the root of the directory

* Update api

* change the file permissions back

Co-authored-by: theofficialgman <28281419+theofficialgman@users.noreply.github.com>

* check gui format version to kill/restart the gui

* only run new gui if old process was killed

prevents running accidentally if there was no gui process active

* move the if check for gui version into the runonce

this worked already but this makes it even more safe. the runonce always succeeds so its hash gets added to the list and skipped in the future.

* remove extra fi

* fix incorrect filepath to app list style

* fix simple bug with source not getting passed to preload script

this caused the preload script to always run and default to the yad format even when it was only supposed to be sourced

* start updates for README

* add some more readme changes

the images should be replaced by someone runing PiOS

* correct a few mistakes in the readme

* add tooltips to the main page

the user might see the default "App folder" tooltip and never check for tooltips again. This adds helpful information to all currently released categories. If a category isn't found, the resulting tooltip will be empty.

* update screenshots with PiXflat GTK theme

* correct readme file link

* Delete genapplist-yad.c

* Update categories

Removed Categories:
Editors

Added Categories:
System Management
Terminals
Programming
Creative Arts
Engineering
Office

New subcategory under tools:
Emulation

* update tooltips for new categories

* New category icons

* update uninstall icon

remove background

* category icon updates

* more category icon improvements

* Add Communication category

* add communication tooltip

* Update README.md

Co-authored-by: Botspot <j.allen.marchand@outlook.com>
Co-authored-by: Botspot <54716352+Botspot@users.noreply.github.com>
Co-authored-by: CleanMachine1 <78213164+CleanMachine1@users.noreply.github.com>
Co-authored-by: Ryan Fortner <71036629+ryanfortner@users.noreply.github.com>