solve errors if user spams the GUI back button when preload is running
if the lists timestamps do not match, the list generation is done at the time the user clicks on the category. previously, each line of the list was directly echoed to the pipe, meaning if the user clicked the back button, they could send data at the same time that other data was still loading in.

change it so that all data is generated and then read the list file into the pipe at the very end.