write to a unique temporary listfile
it can happen that the user interaction triggers a preload list generation at the same time that the preload-daemon is currently  generating that same list. this causes the listfile to have combined contents from both scripts (aka: corruption).

to fix any case of this, use a unique temporary listfile for preload and move that file to the final location when finished. this causes the resulting listfile to always be formatted correctly.