speedup generating LISTS (30% increase in speed)
app_categories is very slow and is a major factor in the preload speed for small categories.

Revert "speedup generating LISTS (30% increase in speed)"

This reverts commit 2f796cfd32b66d3b69642bb4049fd3c1366d1b40.

speedup generating LISTS (30% increase in speed)

using a new api function (so that special "virtual" categories are still accounted for.

for now, obtaining the main page is still done under the old (slower) app_categories