dangerous lunar client update
this uses the files directly from the git since @gl91306 hasn't made a release yet. will switch over to release once it is published. this hotfix prevents users from being left with no install.

unknown if adding the mesa override is needed as he says he fixed that in the appimage itself