add functions list_intersect_partial and list_subtract_partial
pi-apps depends on partial matches for filtering out packages and standard apps when the selected show apps setting is not All.

the recently optimized list_subtract and list_intersect functions did not support this (and they should not, since they are not a match)

this adds back in the old functions under a new name for only this case as all other areas are a direct match.