Fix repo_refresh creating local-packages/source.list, close #1981
We've occasionally had error reports where the pi-apps local repository
was not found in the apt update. Turns out the cause of this issue was
if /etc/apt/sources.list did not have a trailing newline like it usually
does. If this happens, the pi-apps-local-packages line is appended to
the end of the last line (which is usually a commented deb-src line)
Moving pi-apps-local-packages to the first line fixes this problem.
Alternatively, I could have made it place a newline first, but this way
seems cleaner.