users appear to not understand that they need to respond to bug reports
a significant number of bug reports have come in over the years that get no response from the original poster. add a required check telling the user that they are expected to respond to bug reports when they create them. also give them resources for how to write a good bug report