smarter new app detection
If an app was previously only available for one architecture that wasn't the users architecture and was updated to support the users architecture it would not show as a new app. It would be hidden in the compressed app updates.

This adds detection for these cases so that these apps will correctly show as a new app. Useful for when/if the remaining ARM32 ONLY apps get ported to ARM64.