BUGFIX: tail follow descriptor instead of name
tail -F is the same as saying tail --follow=name --retry. this failed to identify the filename on some systems (potentially systems with some different system languages).

instead of following by the filename, use the file descriptor (the default for -f) which works on these systems and for normal users

closes: #2187
closes: #2117
closes: #1842