allow github actions to report actual nproc output
github actions updater runs QEMU which keeps as much compiled code in RAM as possible. This means that in general the free ram is less than 500MB. However, QEMU will reduce its RAM useage as necessary so there is no reason to limit compilations to only one thread. The runners we have free access to only have 2 cores so 7GB of ram even when using QEMU emulation should be plenty.