api: fix will_reinstall in event of install script rename
Old old behavior did not take into account the removal of compatibility
for a particular architecture. If an app first has an install script but
the upgrade only has install-64 script, a 32-bit Pi-Apps install would
incorrectly try to reinstall the app.

Old behavior checked if the local script name existed in the update
folder, and if not found, reinstallation would be disabled. This fixed
the first problem, but broke migrations from install-* to install.

New behavior fixes both issues by determining the script name
individually in each place, allowing for script renames but preventing a
reinstall in cases where compatibility is removed.