Arduino: improve light/dark mode selection window
Previous window showed a combined image of both themes side by side,
with light theme on the left and dark theme on the right. But the choice
buttons listed dark mode on the left and light mode on the right. While
correcting this, I chose to use yad's --form mode this time to compactly
embed both images on each button, as well as mention in tooltips which
one is the default choice.