Updater: lack of internet bugfix
check if update app folder exists before doing anything

it can happen that executing the updater from the pi-apps GUI the update folder is missing. if the user has no internet or internet issues the update folder will be removed due to the failed git pull and then wait in the loop for a new git clone. the GUI updater executes check_repo with the "fast" option so it skips updating the pi-apps update folder and obtains any previously determined updatable apps and files. if we do not check for this then apps will be removed, the older version moved to trash, and then the new version will fail to copy over since it does not exist. this is undesirable behavior