api: optimize release info scripting
lsb_release can be slow to return (up to 0.1seconds). it is faster to call it once and obtain all the information necessary and then map it to the proper variables then to call it multiple times. these changes improve the speed from ~0.8 seconds runtime to ~0.3 seconds runtime. in addition, if the variables are already present, the call is skipped entirely.