wrap manage script in {}, same as the updater
if the new version of a script (with whitespace lines removed) is longer than the original version of a script (with whitespace lines removed) then bash will attempt to execute the new lines that follow in the new script when the original script reaches its end. this often produces syntax errors and undesirable affects. simply explicitly exit the script when the script has finished to avoid this issue.

if the new script is shorter or equal in length to the original script then this is not an issue which made it harder to catch. users who update will not be affected by the bug in the original manage script since the manage file gets updated while manage is not running in the old system. the new system updates manage while manage is running so it needs this fix.