add `terminal_manage_multi` function
have `terminal_manage` function as a legacy wrapper to `terminal_manage_multi`. this allows for pi-apps scripts (even old ones) to still call `terminal_manage` with two arguments. this is very important for the initial update where the old GUI is still running and the new manage script will be called.