remove `write_list` from end of daemon loop
avoid writing the queue to the manage daemon GUI after an install/uninstall/update/refresh has finished because the queue variable may be out of date. if I user added a new app to the queue while the daemon was busy performing an action on an app, then the new app the user added would disappear and then quickly re-appear when the loop was re-entered.