manage: bugfix, don't spawn second terminal for "app-script doesn't match version in update folder" dialog
the "app-script doesn't match version in update folder" dialog spawns when a user goes to install an app that has available updates. this is part of the `validate_apps_gui`. previously the PID for the master manage daemon was not written to the pid file until after `validate_apps_gui`, causing `terminal_manage_multi` to spawn a new terminal since the pid file is missing when it gets called.

this commit changes it so that the pid file is written before `validate_apps_gui`. An additional check is added so that the file gets properly removed if all selections get invalidated such that the queue is empty. in addition, if the pid file exists but the process for that PID does not exist, the PID of the current instance is written to the file just like before.