gui: kill tail when yad exits
GNU coreutils 8.30 has a bug where tail will not exit if the piped command exits under certain scenarios. this is fixed in later versions of GNU coreutils. This would cause the settings button and search buttons to not work in the gui and tail to infinite loop and 100% cpu.

This bug was known to affect Ubuntu Focal with GNU coreutils 8.30 with the default gnome desktop.