Minecraft Bedrock: remove unnecessary dependencies and bump version
the bedrock launcher keeps itself up to date which is why the version was not bumped before. Since making the change to remove an unneeded dependency (libgl1-mesa-dri already depends on the correct version of llvm) triggers a pi-apps update, might as well increase the launcher version to latest while we are at it.