WorldPainter: make .desktop location more general
`xdg-desktop-menu` is used to install the .desktop file as part of the applications postinst script. the director which it installs to is different per system. On piOS it appears that one possible location is `fkms` while on ubuntu it is `ubuntu`. There are likely many locations so use a `*` to account for all of them.