Wine (x86/x64): remove multiple dlls
wine has builtin support for every version of DX and wine support and performance is often better and more bug free than using these dlls. remove them for a better out of the box experience and align with WineHQ https://wiki.winehq.org/FAQ

also remove dotnet20 as it hangs on installation

dotnet20 installs on an x86_64 PC with WOW64 so this is likely a box64 issue.