gui: remove while loop and correct app_search_gui function
yad now calls the `run_user_selection` function directly which allows for DEs to properly handle window focusing. also modify the parsing of yad output in app_search_gui since that strangely changed as a side affect of this change

workaround array variables not being passed to `bash -c`

bash -c specifically does not pass the contents of array variables. we use an array variable to set the default yadflags and it is the only variable that need to be set that we don't have access to inside bash -c

without this the spawned yad windows from app_search_gui, the details window, and updater will be missing the default flags

this also affects the settings yad GUI which spawns other scripts with bash -c, however all of those scripts don't rely on the yadflags (or re-set them internally) so the issue doesn't occur there