use a pipe to store list of detailspid
`kill_details` is a blocking function and part of it is sent to the background so as to not block further execution. To prevent accidentally reading the new detailspid that gets added to the list, make sure to read the current contents first in a blocking manner and then send the kill loop to the background.