Improve update-when-installing experience
Now choosing to update an app before installing it will directly update
only that app and the api script, instead of running the full updater
script. A user chose to close out of updater when the app was not
prominently shown as updatable (because it was in the list of
refreshable apps), and this method makes more sense anyway. api script
is updated in case the app script needs a new api function.