Github Desktop: update to official Linux port ARMhf/ARM64 debs
upstream now builds for ARMhf/ARM64 through cross compilation in a bionic docker container on github actions. This allows for very old versions of linux to remain supported on all architectures (x64, armhf, arm64). This is the completion of a months long effort to make upstream and its dependencies compatible with ARMhf/ARM64.