api: `log_diagnose` don't attempt writes to logfile by default
only write if argument "allowwrite" is passed. this restores the function to what it was before we added additional diagnosis that wrote to the logfile. it is not ideal to always write to the logfile since users may want to run log_diagnose themselves and won't expect their logfile to change after doing so.