correct ca7a7cd75b71bb87d6e812f18f19a1770b4c05bd to actually update users .desktop files
the runonce previously checked for `StartupWMClass` to exist in the .desktop files before updating the settings and updater .desktop files. This caused updating these files to be skipped on users that already had this change from last year. This check is unnecessary as this is already a runonce and won't run again unless modified.