api: add `User error (reporting allowed):` designation for scripts to use
there are cases where we known an error is a user error and want to self diagnose an issue for the user but at the same time still want them to be able to send in logs. Such cases might include seeing how common an issue is or finding new issues where previously not thought possible.