Firefox Rapid Release: only pin firefox to 1000
thunderbird is available from the same PPA while potentially being a lower version than available in Ubuntu repos which still have it as a true deb package. 

We would either need to mark thunderbird from Ubuntu as pin priority zero or, as is done in this commit, only mark firefox packages as 1000 from this repo.