api: add `deb-multimedia.org` check to `is_supported_system`
`deb-multimedia.org` is a common enough third party repo that has specific versions per each debian release. mixing and matching versions is not supported WILL result in a broken apt system.

Additionally, they support "stable" "oldstable" and "oldoldstable" target names in addition to the traditional codename (bookworm/bullseye/buster/etc), these "stable" based targets should never every be used in any circumstance as they will change arbitrarily when the repo maintainer updates to the newest debian release naming.

the system will be marked as unsupported in the case that the user uses any of the "stable" based names or a debian codename which does not match the system's codename.