Better Chromium: support more folder paths
the `customizations` folder is a ubuntu concept (note that rpiOS chromium is derived from the original ubuntu chromium packaging). the folder name is defined in the chromium wrapper script as a function of the binary name. ubuntu packages chromium as `chromium-browser` however upstream chromium packaging refers to it as `chromium-browser-$channel` with channel being stable, beta, or unstable