remove 3 months out of date updater
did not solve any problems that were not already handled by the GUI script mission critical re-download functionality

additionally, when the 3-months-updater runs as part of a update (one of the most common cases), all queued updates fail (each with a prompt about them failing) leading to a very poor user experience. the number of errors can be so large that it crashes piOS wayfire to the login screen

any future attempt at force updating should be handled through the updater script, not manage. the updater and api scripts are always guaranteed to be functional (as long as the repo version is functional)