api: refactor `remove_repofile_if_unused` to use `anything_installed_from_uri_suite_component`
handles .sources files and .list files with multiple components and multiple suites (as allowed) and filters out disabled sources.

in deb822, it is convention to follow a field separator colon by a space and then the package name but the specification allows for no space or any number of spaces or tab characters. trim all tab and space characters to a single space character for components and suites that can have multiple strings of output