`manage`: don't run is_supported_system check for daemon mode
the previous setup caused is_supported_system to run every time when a user added a install/uninstall to the manage daemon even though that manage instance exited once it was added to the queue and the supported yes/no status went unused. this causes a noticeable delay 3-10 seconds in testing on adding install/uninstalls to the queue.

this change fixes that to only run is_supported_system when needed on the manage instance acting as the daemon