`manage`: move yadlist creation before is_supported_system check
the pid and yadlist need to be created ASAP after it is decided that the current terminal will be acting as the daemon. if the yadlist is created too late the user could overwrite it with install/uninstall names before the main daemon has a chance to create it as a pipe.