api: improve https://github.com/Botspot/pi-apps/commit/368f6e3f040a40ecd88e14f8cee4215d52e9d5f4
before this commit, only a single package was attempted to dry-run install which will not result in a "breaks" output from apt but instead diffferent "depends" output with "but it is not going to be installed". this is different from the logging output which attempts to install multiple packages at one time.

after this commit, the debug output itself (where multiple packages are attempted to be installed at once) is used for parsing. this also speeds up execution greatly.