api: update `shlink_link` to add OS and Hardware information to click user agent (#2648)
The machine-id and serial-number are on-way hashed and cannot be reversed.

Also unify all click commands to use `shlink_link`. Modify `shlink_link` to check for the presence of `No` in the `Enable analytics` preference file to disable clicks instead of the absence of `Yes`.

Also make `shlink_link` return immediately and fork the process.