api: replace here-string in `log_diagnose` with process substitution
a here-string creates a temporary file on disk which can lead to failure to process all greps in the case of disk space exhaustion. a here-string is also subject to max argument length exhaustion on some systems/kernels leading to failure to process all greps.

process substitution occurs in memory and is not subject to the same limitations.