Jetson Category Overrides: add Steam Link
even though Steam Link is a package app and would be hidden already, we use this file in the `update-wiki.sh` script for the website to generate webpages. Without this the "Steam Link" page is generated for Nintendo Switch and Nvidia Jetson. Unfortunately this webpage has gotten traction in the past month and is confusing users.