only install `software-properties-common` dependency when available
the `software-properties-common` package contains the program `add-apt-repository` which is used in `ubuntu_ppa_installer` in pi-apps to assist with installing PPAs on Ubuntu (and derivative) systems. `software-properties-common` has been removed on Debian Trixie (13). As of now, Ubuntu has no plans on removing this package since Ubuntu developed it and uses it for managing PPAs and 3rd party repositories.

closes https://github.com/Botspot/pi-apps/issues/2798